Add new connection error notifier and error field to union
The user provided connection notifier was called on startup only in case
of an successfully established connection or if an error happened while
connecting. This patch adds a new (internal) client connection error
notifier that is called from the ISelectClient `error` method, in that
case the client Connection class. That happens when the ISelectClient is
unregistered from epoll due to a exception thrown in the RequestOnConn
fiber.

The exception is passed to the user by utilizing a new error field in
the ConnNotificationUnion (defined in the ConnectionSet).